djangoreactnativeauth2/djangoauthapi1/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==1.1.0
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
djangorestframework-simplejwt==5.2.0
Pillow==9.1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
